+++ cmd/main.go
@@ -79,7 +79,6 @@ const (
 	flagManagementClusterRegion       = "management-cluster-region"
 
 	// Monitoring configuration flag names
-	flagAlertmanagerSecretName                = "alertmanager-secret-name"
 	flagAlertmanagerURL                       = "alertmanager-url"
 	flagMonitoringEnabled                     = "monitoring-enabled"
 	flagMonitoringNetworkEnabled              = "monitoring-enable-network-monitoring"
@@ -254,8 +253,6 @@ func parseFlags() (err error) {
 		"The region of the management cluster.")
 
 	// Monitoring configuration flags
-	pflag.StringVar(&cfg.Monitoring.AlertmanagerSecretName, flagAlertmanagerSecretName, "",
-		"The name of the secret containing the Alertmanager configuration.")
 	pflag.StringVar(&cfg.Monitoring.AlertmanagerURL, flagAlertmanagerURL, "",
 		"The URL of the Alertmanager API.")
 	pflag.BoolVar(&cfg.Monitoring.Enabled, flagMonitoringEnabled, false,